Brings the v0.3.3 line to master. The milestone's headline outcome is a
library that is provably green across the full supported toolchain matrix,
whose test and benchmark gates assert only what each path can honestly claim,
and whose documentation and examples front door is ready for external adopters.
The determinism slice of the RT theme landed; RT/embedded allocation-freedom
and the on-device proof carry forward to the next line.

Portability & CI:
- Compile the library and tests clean under the strict -Werror warning gate.
- Green across clang-18, gcc-14, MSVC, Apple-clang (Xcode 15.4/16.2), and both
the C++20 floor and the C++23 line, plus the system-Eigen integration path.
- Resolve cross-compiler breakages: MSVC bigobj/M_PI and unqualified
argmin::unexpected, gcc unsigned-compare and -Wsign-conversion, pre-P2113
clang CTAD deduction, old Apple-clang aggregate CTAD, and the GCC-14
TEMPLATE_TEST_CASE_SIG quirk.
- Drop a deprecated volatile compound assignment from the exceptions-off probe;
gate the warnings job and silence a spurious clang-18 driver warning.

Sanitizers hardened to gates:
- Promote MemorySanitizer to a hard gate, built against an instrumented libc++
with an uninstrumented unwinder so its reports are real signal.
- Define the ccsa reject-path inequality buffer so it is not read uninitialized.

Determinism & numerical honesty:
- Assert the projected Gauss-Newton determinism claims each path can honestly
make: run-to-run and post-reset bit-exactness stay exact; cross-instantiation
numerics move to a measured tolerance rather than a false bit-identity.
- Hold pinned coordinates exactly in the trust-region restart projector.
- Derive the CMA-ES Marsaglia uniforms portably from the raw engine bits.
- Keep the dense_ldl_bfgs SPD stress well-conditioned and independent of the
standard-library RNG stream so LLT is not platform-fragile.
- Sharpen the HS071 optimum to full double precision and measure its
fast/accurate wall ratio with a jitter-robust estimator.

Benchmark & publication gate:
- Wire an independent oracle into the recurring publish gate, witness pass-cell
correctness against it, and classify disposition flips as corrected truth
fields rather than regressions.
- Gate per-iteration instruction cost by ratio and a real-time ceiling.
- Regenerate the publication regression baseline from a multi-seed release run.

Adopter front door:
- Reframe the README as an adopter front door; add getting-started and
publication docs, a CONTRIBUTING guide, and a Codecov coverage badge.
- Add a runnable examples tree (constrained SLSQP, least-squares LM,
derivative-free BOBYQA, step-level control) built in CI.

Hygiene: 
- Purge planning-artifact identifiers from headers, tests, benchmarks, and docs.
- Correct and complete the academic bibliography and in-code citations, and
remove a fabricated theorem attribution.

+++ lib/argmin/include/argmin/test_functions/ik_pose_batch.h
@@ -70,13 +70,13 @@ struct ik_pose_batch
     // singularities at the chosen home/target configurations). Link
     // lengths and offsets are dimensionless; the workspace spans a
     // sphere of radius ~ sum(|a_i|) + sum(|d_i|) about the base.
-    [[nodiscard]] static std::array<dh_row, NJoints> dh_table() noexcept
+    [[nodiscard]] static std::array<dh_row, static_cast<std::size_t>(NJoints)> dh_table() noexcept
     {
         static_assert(NJoints == 6,
                       "ik_pose_batch currently ships a 6-DOF DH table; "
                       "templating on NJoints is reserved for future "
                       "kinematic shapes.");
-        std::array<dh_row, NJoints> table{};
+        std::array<dh_row, static_cast<std::size_t>(NJoints)> table{};
         constexpr Scalar pi = std::numbers::pi_v<Scalar>;
         table[0] = dh_row{Scalar(0),    pi / Scalar(2),  Scalar(1.0),  Scalar(0)};
         table[1] = dh_row{Scalar(1.0),  Scalar(0),       Scalar(0),    Scalar(0)};
@@ -90,9 +90,9 @@ struct ik_pose_batch
     // K target positions, distributed on a circle in the xy plane at
     // z = 0.8 with radius 1.2 about the base. All inside the reachable
     // workspace (sum of link lengths is approximately 2.8).
-    [[nodiscard]] static std::array<Eigen::Matrix<Scalar, 3, 1>, K> targets() noexcept
+    [[nodiscard]] static std::array<Eigen::Matrix<Scalar, 3, 1>, static_cast<std::size_t>(K)> targets() noexcept
     {
-        std::array<Eigen::Matrix<Scalar, 3, 1>, K> p_targets{};
+        std::array<Eigen::Matrix<Scalar, 3, 1>, static_cast<std::size_t>(K)> p_targets{};
         constexpr Scalar pi = std::numbers::pi_v<Scalar>;
         constexpr Scalar radius = Scalar(1.2);
         constexpr Scalar z_target = Scalar(0.8);
@@ -169,8 +169,8 @@ struct ik_pose_batch
     // Returns the end-effector transform cumulative[NJoints].
     [[nodiscard]] static Eigen::Matrix<Scalar, 4, 4> forward_kinematics(
         const Eigen::Matrix<Scalar, NJoints, 1>& theta,
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints + 1>& cumulative,
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints>& pre_joint) noexcept
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints) + 1>& cumulative,
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints)>& pre_joint) noexcept
     {
         const auto table = dh_table();
         cumulative[0].setIdentity();
@@ -191,8 +191,8 @@ struct ik_pose_batch
     // pre_joint[i] frame; that axis and origin feed the cross product
     //   J_p.col(i) = z_axis_i x (p_ee - p_pre_i).
     [[nodiscard]] static Eigen::Matrix<Scalar, 3, NJoints> position_jacobian(
-        const std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints + 1>& cumulative,
-        const std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints>& pre_joint) noexcept
+        const std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints) + 1>& cumulative,
+        const std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints)>& pre_joint) noexcept
     {
         const Eigen::Matrix<Scalar, 3, 1> p_ee =
             cumulative[NJoints].template block<3, 1>(0, 3);
@@ -222,8 +222,8 @@ struct ik_pose_batch
     {
         const auto p_targets = targets();
         Scalar acc = Scalar(0);
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints + 1> cumulative;
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints> pre_joint;
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints) + 1> cumulative;
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints)> pre_joint;
         Eigen::Matrix<Scalar, NJoints, 1> theta_k;
         for(int k = 0; k < K; ++k)
         {
@@ -245,8 +245,8 @@ struct ik_pose_batch
     {
         const auto p_targets = targets();
         g.setZero();
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints + 1> cumulative;
-        std::array<Eigen::Matrix<Scalar, 4, 4>, NJoints> pre_joint;
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints) + 1> cumulative;
+        std::array<Eigen::Matrix<Scalar, 4, 4>, static_cast<std::size_t>(NJoints)> pre_joint;
         Eigen::Matrix<Scalar, NJoints, 1> theta_k;
         for(int k = 0; k < K; ++k)
         {
